+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [BUGFIX] Fix issues with kubernetes integration when were setting a global multisite setting it was not applied to the services
 - [UI] Update web UI setup wizard to handle when a reverse proxy already exists but no admin user is configured
 - [UI] Fix issues with multiple settings on the global_config not being able to be deleted in specific cases
+- [AUTOCONF] Fix issues with globally set settings overridden by default values not being saved correctly in database
 - [SECURITY] Update security headers in default pages and error pages for improved security
 - [DEPS] Updated LuaJIT version to v2.1-20240626
 - [DEPS] Updated coreruleset-v4 version to v4.4.0